pyproject.toml
[tool.poetry]
name = "Site Validator"
version = "2.1.0"
description = "Site validator"
authors = ["bwg.it425 <bwg.it425@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.88.0"
uvicorn = "^0.20.0"
pydantic = {extras = ["dotenv", "email"], version = "^1.9.0"}
dependency-injector = "^4.40.0"
yoyo-migrations = "^8.1.0"
aiopg = "^1.3.3"
bcrypt = "^4.0.1"
python-jose = {version = "3.3.0", extras = ["cryptography"]}
pydantic-yaml = "^0.8.0"
aiohttp = "^3.8.2"
pyotp = "^2.7.0"
httpx = "^0.23.0"
aiosqlite = "^0.17.0"
httpcore = "^0.16.2"
pyyaml = "^6.0"
selenium = "^4.8.0"
xmltodict = "^0.13.0"
celery = {extras = ["redis"], version = "^5.2.7"}
flower = "^1.2.0"
pytesseract = "^0.3.10"
requests = "^2.28.2"
selenium-wire = "^5.1.0"
undetected-chromedriver = "^3.4.6"

[tool.poetry.group.dev.dependencies]
MarkupSafe = "^2.1.0"

[tool.poetry.group.formatters.dependencies]
autoflake = "^1.4"
isort = "^5.10.1"
black = "^22.1.0"
docformatter = "^1.4"
add-trailing-comma = "^2.2.1"

[tool.poetry.group.sasts.dependencies]
itsdangerous = "^2.1.2"
safety = "^2.3.3"
bandit = "^1.7.1"

[tool.poetry.group.linters.dependencies]
flake8 = "^4.0.1"
mypy = "^0.931"

[tool.poetry.group.tests.dependencies]
pytest-asyncio = "^0.20.2"
coverage = "^6.3.2"
pytest-cov = "^4.0.0"

[tool.poetry.group.docs.dependencies]
Sphinx = "^5.3.0"
sphinx-pydantic = "^0.1.1"
sphinx-rtd-theme = "^1.1.1"



[tool.poetry.group.testing.dependencies]
locust = "^2.14.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
